+++ src/sc_linac_physics/__init__.py
@@ -1,3 +1,3 @@
 """SC Linac Physics package."""
 
-__version__ = "8.1.0"
+__version__ = "8.3.0"